Desktop/music_player/requirements.txt
pygame==2.6.1
pydub==0.25.1
numpy==1.24.3
Pillow==10.0.1
